feat: 미션 카테고리 확장 및 UI 개선
- MissionCategory에 category/type 조합 적용 (걸음 수 챌린지, 출석 챌린지 구분)
- CategoryChip 클릭 상태별 스타일링 (배경색, 테두리, 텍스트색 변경)
- WalkingViewModel ViewModel 공유 문제 해결 (WalkingGraph 범위 공유)
- 임시 세션(userId=0) 생성 문제 해결 (userId 기반 세션 생성)
- 시간대 통일 (모든 시간 관련 함수 UTC 사용)
- MissionCard 첫 번째 카드 특별 처리 (isActive 파라미터 추가)
- 더미 데이터 생성 시 userId 설정
- Lottie 리소스 통합 및 최적화
- ProfileImageState 개선 및 UI 상태 관리 향상

+++ app/src/main/java/team/swyp/sdu/data/remote/user/dto/FriendListItemDto.kt
@@ -12,11 +12,12 @@ data class FriendListItemDto(
     val nickname: String,
     
     @SerialName("userId")
-    val userId: Int,
+    val userId: Long,
     
     @SerialName("imageName")
     val userImageUrl: String? = null,
 )
+
